xAI: restore generic auth and x_search seams
