+++ extensions/xai/api.ts
@@ -14,17 +14,7 @@ export {
   XAI_DEFAULT_MAX_TOKENS,
 } from "./model-definitions.js";
 export { isModernXaiModel, resolveXaiForwardCompatModel } from "./provider-models.js";
-export {
-  __testing as xSearchTesting,
-  buildXaiXSearchPayload,
-  requestXaiXSearch,
-  resolveXaiXSearchInlineCitations,
-  resolveXaiXSearchMaxTurns,
-  resolveXaiXSearchModel,
-  type XaiXSearchOptions,
-} from "./src/x-search-shared.js";
-import { normalizeXaiModelId } from "./model-id.js";
-export { normalizeXaiModelId };
+export { normalizeXaiModelId } from "./model-id.js";
 
 export const XAI_TOOL_SCHEMA_PROFILE = "xai";
 export const HTML_ENTITY_TOOL_CALL_ARGUMENTS_ENCODING = "html-entities";
