+++ extensions/xai/x-search.ts
@@ -1,0 +1,8 @@
+export {
+  buildXaiXSearchPayload,
+  requestXaiXSearch,
+  resolveXaiXSearchInlineCitations,
+  resolveXaiXSearchMaxTurns,
+  resolveXaiXSearchModel,
+  type XaiXSearchOptions,
+} from "./src/x-search-shared.js";
